correction entre product-card et product-list
