+++ src/app/app.module.ts
@@ -5,9 +5,9 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './components/header/header.component';
 import { FilterByTagComponent } from './components/filter-by-tag/filter-by-tag.component';
-import { ProductListComponent } from './components/product-list/product-list.component';
+import { CentimesToEurosPipe, ProductListComponent } from './components/product-list/product-list.component';
 import { AllProductsComponent } from './pages/all-products/all-products.component';
-import { CentimesToEurosPipe, ProductCardsComponent } from './components/product-cards/product-cards.component';
+import { ProductCardsComponent } from './components/product-cards/product-cards.component';
 
 
 
@@ -21,7 +21,7 @@ import { CentimesToEurosPipe, ProductCardsComponent } from './components/product
     CentimesToEurosPipe,
     HeaderComponent,
     FilterByTagComponent,
-    
+
   ],
   imports: [
     BrowserModule,
